requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.1.5
django-environ==0.9.0
django-soft-delete==0.9.21
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
